% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_weight.R
\name{plot_weight}
\alias{plot_weight}
\title{Plot the weight distribution of the optimal design}
\usage{
plot_weight(u, w_hat, S)
}
\arguments{
\item{u}{The discretized design space}

\item{w_hat}{The estimated weight of each design point}

\item{S}{The design space}
}
\value{
The plot that shows the given optimal design
}
\description{
Plot the weight distribution of the optimal design
}
\details{
TODO
}
