% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Aopt.R
\name{Aopt}
\alias{Aopt}
\title{Calculate the A-optimal design under the SLSE}
\usage{
Aopt(N, u, tt, FUN, theta)
}
\arguments{
\item{N}{The number of sample points in the design space}

\item{u}{The discretized design space}

\item{tt}{The level of skewness}

\item{FUN}{The function to calculate the derivative of the given model.}

\item{theta}{The parameter value of the model}
}
\value{
CVX object that contains the solved optimal design
}
\description{
Calculate the A-optimal design under the SLSE
}
\details{
TODO
}
