% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_phiA.R
\name{calc_phiA}
\alias{calc_phiA}
\title{Calculate the loss function of the A-optimal design}
\usage{
calc_phiA(N, u, w_hat, theta, FUN, tt, A)
}
\arguments{
\item{N}{The number of sample points in the design space}

\item{u}{The discretized design space}

\item{w_hat}{The weight of each design points}

\item{theta}{The parameter value of the model}

\item{FUN}{The function to calculate the derivative of the given model.}

\item{tt}{The level of skewness}

\item{A}{The calculated covariance matrix}
}
\value{
The loss of the model at each design points
}
\description{
Calculate the loss function of the A-optimal design
}
\details{
TODO
}
