% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_phiD.R
\name{calc_phiD}
\alias{calc_phiD}
\title{Calculate the loss function of the D-optimal design}
\usage{
calc_phiD(N, u, w_hat, FUN, tt, A, theta)
}
\arguments{
\item{N}{The number of sample points in the design space}

\item{u}{The discretized design space}

\item{w_hat}{The weight of each design points}

\item{FUN}{The function to calculate the derivative of the given model.}

\item{tt}{The level of skewness}

\item{A}{The calculated covariance matrix}

\item{theta}{The parameter value of the model}
}
\value{
The loss of the model at each design points
}
\description{
Calculate the loss function of the D-optimal design
}
\details{
TODO
}
