% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dopt.R
\name{Dopt}
\alias{Dopt}
\title{Calculate the D-optimal design under the SLSE}
\usage{
Dopt(N, u, tt, FUN, theta, num_iter = 2500)
}
\arguments{
\item{N}{The number of sample points in the design space}

\item{u}{The discretized design space}

\item{tt}{The level of skewness}

\item{FUN}{The function to calculate the derivative of the given model.}

\item{theta}{The parameter value of the model}

\item{num_iter}{Maximum number of iteration}
}
\value{
A list that contains 1. Value of the objective function at solution. 2. Status. 3. Optimal design
}
\description{
Calculate the D-optimal design under the SLSE
}
\details{
This function calculates the loss function of the design problem under the D-optimality. The loss function under D-optimality is defined as the log determinant of the inverse of the Fisher information matrix
}
