% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_direction_Aopt.R
\name{plot_direction_Aopt}
\alias{plot_direction_Aopt}
\title{Calculate the loss function of the A-optimal design}
\usage{
plot_direction_Aopt(u, w_hat, tt, FUN, A, phi, theta, q, S, N)
}
\arguments{
\item{u}{The discretized design space}

\item{w_hat}{The weight of each design points}

\item{tt}{The level of skewness}

\item{FUN}{The function to calculate the derivative of the given model.}

\item{A}{The calculated covariance matrix}

\item{phi}{The loss loss function on each design point}

\item{theta}{The parameter value of the model}

\item{q}{The penality term}

\item{S}{The design space}

\item{N}{The number of sample points in the design space}
}
\value{
The loss of the model at each design points
}
\description{
Calculate the loss function of the A-optimal design
}
\details{
TODO
}
