% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_direction_Dopt.R
\name{plot_direction_Dopt}
\alias{plot_direction_Dopt}
\title{Calculate the loss function of the D-optimal design}
\usage{
plot_direction_Dopt(design, tt, FUN, A, phi, theta)
}
\arguments{
\item{design}{The resulted design that contains the design points and the associated weights}

\item{tt}{The level of skewness}

\item{FUN}{The function to calculate the derivative of the given model.}

\item{A}{The calculated covariance matrix}

\item{phi}{The loss loss function on each design point}

\item{theta}{The parameter value of the model}
}
\value{
The plot of the directional derivative of a D-optimal design
}
\description{
Calculate the loss function of the D-optimal design
}
\details{
This function produces the figure for the directional derivative of the given D-optimal design of the compact supports. According to the general equivalence theorem, for an optimal design, all the directional derivative should be below zero line.
}
