% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_weight.R
\name{plot_weight}
\alias{plot_weight}
\title{Plot the weight distribution of the optimal design}
\usage{
plot_weight(design)
}
\arguments{
\item{design}{The resulted design that contains the design points and the associated weights}
}
\value{
The plot that shows the given optimal design
}
\description{
Plot the weight distribution of the optimal design
}
\details{
TODO
}
